[MERGE] Talita

+++ app/app/src/main/java/listbuy/me/listbuy/lista/Sincronizacoes/SincronizaLogin.java
@@ -29,9 +29,6 @@ import listbuy.me.listbuy.Welcome;
 import listbuy.me.listbuy.lista.DbConn;
 import listbuy.me.listbuy.lista.Lista_inicial;
 
-/**
- * Created by Talitadossantoscastr on 13/11/2016.
- */
 
 public class SincronizaLogin extends AsyncTask<String, String, String> {
 
@@ -118,7 +115,7 @@ public class SincronizaLogin extends AsyncTask<String, String, String> {
             if (message.equalsIgnoreCase("success")) {
                 String dados = api_result.getString("object");
                 JSONObject dados_result = new JSONObject(dados);
-                int id_consumidor = dados_result.getInt("id_consumidor");
+                Long id_consumidor = dados_result.getLong("id_consumidor");
                 String nome = dados_result.getString("nome");
                 String email = dados_result.getString("email");
                 String senha = dados_result.getString("senha");
