+++ backend/src/main/java/br/puccamp/listbuy/entities/ListaPessoas.java
@@ -6,7 +6,7 @@ package br.puccamp.listbuy.entities;
 public class ListaPessoas {
 
     private Listas listas;
-    private Integer id_consumidor;
+    private Long id_consumidor;
 
     public Listas getListas() {
         return listas;
@@ -16,11 +16,11 @@ public class ListaPessoas {
         this.listas = listas;
     }
 
-    public Integer getId_consumidor() {
+    public Long getId_consumidor() {
         return id_consumidor;
     }
 
-    public void setId_consumidor(Integer id_consumidor) {
+    public void setId_consumidor(Long id_consumidor) {
         this.id_consumidor = id_consumidor;
     }
 }
