+++ backend/src/main/java/br/puccamp/listbuy/entities/ProdutoCadastrado.java
@@ -7,15 +7,15 @@ import java.sql.Date;
  */
 
 public class ProdutoCadastrado {
-    private Integer ID_CONSUMIDOR;
+    private Long ID_CONSUMIDOR;
     private Produtos PRODUTO;
     private Date DATA_ICS;
 
-    public Integer getID_CONSUMIDOR() {
+    public Long getID_CONSUMIDOR() {
         return ID_CONSUMIDOR;
     }
 
-    public void setID_CONSUMIDOR(Integer ID_CONSUMIDOR) {
+    public void setID_CONSUMIDOR(Long ID_CONSUMIDOR) {
         this.ID_CONSUMIDOR = ID_CONSUMIDOR;
     }
